+++ examples/src/main/java/org/conceptoriented/bistro/examples/Example2.java
@@ -72,12 +72,15 @@ public class Example2 {
         Object value;
 
         value = itemsAmount.getValue(5); // value = 270.0 = 15 * 18
+        if(Math.abs((double)value - 270.0) > 1e-10) System.out.println(">>> UNEXPECTED RESULT.");
         value = itemsAmount.getValue(21); // value = 450.0 = 25 * 18
+        if(Math.abs((double)value - 450.0) > 1e-10) System.out.println(">>> UNEXPECTED RESULT.");
 
         value = itemsProduct.getValue(5); // value = 0
         value = itemsProduct.getValue(21); // value = 0
 
         value = productsAmount.getValue(0); // value = 720.0 = 270.0 + 450.0
+        if(Math.abs((double)value - 720.0) > 1e-10) System.out.println(">>> UNEXPECTED RESULT.");
     }
 
 }