+++ examples/src/main/java/org/conceptoriented/bistro/examples/Example7.java
@@ -43,6 +43,15 @@ public class Example7 {
         // Volume Weighted Average Price - VWAP = SUM(Price*Volume) / SUM(Volume)
         //
 
+        Column priceVolumeSum = schema.createColumn("PriceSum", quotes, columnType);
+        priceVolumeSum.setDefaultValue(0.0); // It will be used as an initial value
+        priceVolumeSum.roll(
+                time_seconds, // Time stamp
+                60, 0, // 3600 seconds moving average
+                (a,d,p) -> (double)a + (Double.valueOf((String)p[0]) * Double.valueOf((String)p[1])), // [out] + [Price] * [Amount]
+                quotes.getColumn("Price"), quotes.getColumn("Amount")
+        );
+
         Column volumeSum = schema.createColumn("VolumneSum", quotes, columnType);
         volumeSum.setDefaultValue(0.0); // It will be used as an initial value
         volumeSum.roll(
@@ -52,19 +61,10 @@ public class Example7 {
                 quotes.getColumn("Amount")
         );
 
-        Column priceSum = schema.createColumn("PriceSum", quotes, columnType);
-        priceSum.setDefaultValue(0.0); // It will be used as an initial value
-        priceSum.roll(
-                time_seconds, // Time stamp
-                60, 0, // 3600 seconds moving average
-                (a,d,p) -> (double)a + (Double.valueOf((String)p[0]) * Double.valueOf((String)p[1])), // [out] + [Price] * [Amount]
-                quotes.getColumn("Price"), quotes.getColumn("Amount")
-        );
-
         Column VWAP = schema.createColumn("VWAP", quotes, columnType);
         VWAP.calc(
                 p -> (double)p[0] / (double)p[1],
-                priceSum, volumeSum
+                priceVolumeSum, volumeSum
         );
 
         //
@@ -76,13 +76,12 @@ public class Example7 {
         Object value;
 
         value = volumeSum.getValue(3); // value = 0,54029262 (3 elements including this one)
-        if(Math.abs((double)value - 0.54029262) > 1e-10) System.out.println("UNEXPECTED RESULT.");
+        if(Math.abs((double)value - 0.54029262) > 1e-10) System.out.println(">>> UNEXPECTED RESULT.");
 
-        value = priceSum.getValue(3); // value = 214,430001602
-        if(Math.abs((double)value - 214.430001602) > 1e-10) System.out.println("UNEXPECTED RESULT.");
+        value = priceVolumeSum.getValue(3); // value = 214,430001602
+        if(Math.abs((double)value - 214.430001602) > 1e-10) System.out.println(">>> UNEXPECTED RESULT.");
 
         value = VWAP.getValue(3); // value = 396,87753203440017374288769667074
-        if(Math.abs((double)value - 396.87753203440017374288769667074) > 1e-10) System.out.println("UNEXPECTED RESULT.");
+        if(Math.abs((double)value - 396.87753203440017374288769667074) > 1e-10) System.out.println(">>> UNEXPECTED RESULT.");
     }
-
 }